rename `RequestContainer` to `MultiRequest`
- update examples in README
- add support for `MultiRequest` to Laravel Service Provider
- fix code style
- remove unused import
- fixes from php-cs-fixer
- update Composer requirements
  - added PHP 7.4
  - update versions for dev requirements
- update change log
